+++ lib/dradis/plugins/projects/upload/v1/template.rb
@@ -38,6 +38,11 @@ module Dradis::Plugins::Projects::Upload::V1
 
       private
 
+      # No-op here, overwritten in V2
+      def create_comments(commentable, xml_comments)
+        true
+      end
+
       def create_activities(trackable, xml_trackable)
         xml_trackable.xpath('activities/activity').each do |xml_activity|
           # if 'validate_and_save(activity)' returns false, it needs
@@ -70,6 +75,8 @@ module Dradis::Plugins::Projects::Upload::V1
         return false unless validate_and_save(issue)
 
         return false unless create_activities(issue, xml_issue)
+
+        return false unless create_comments(issue, xml_issue.xpath('comments/comment'))
 
         true
       end
@@ -333,6 +340,7 @@ module Dradis::Plugins::Projects::Upload::V1
             end
 
             raise "Couldn't create activities for Note ##{note.id}" unless create_activities(note, xml_note)
+            raise "Couldn't create comments for Note ##{note.id}" unless create_comments(note, xml_note.xpath('comments/comment'))
 
             logger.info { "\tNew note added." }
           end