+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## Dradis Framework 3.11 (XXX, 2018) ##
+
+*  Note comments in export/import 
+
 ## Dradis Framework 3.10 (August, 2018) ##
 
 *   Use project scopes
 
 *   Check project existence for default user id
 
-*   Comments export import
+*   Issue comments in export/import
 
 *   Replace Node methods that are now Project methods